refactor: 在模块间迁移部分代码 并且调整 `TestManager` 的注册完成回调实现
- 迁移部分代码从模块 `simulated-player` 到 `test`
- `TestManager` 现在通过 ready 属性（一个 `Promise`）等待完成，`ready` 会携带 `test` 对象兑现

明确模块边界
