+++ src/core/test/manager.ts
@@ -1,11 +1,17 @@
-import { world, type Vector3 } from "@minecraft/server";
+import { system, world, type Vector3 } from "@minecraft/server";
 import { register, type Test } from "@minecraft/server-gametest";
+
+const overworld = world.getDimension('overworld');
 
 export class TestManager {
     private readonly maxTicks = 20 * 60 * 60 * 24 * 365;
     private _testLocation: Vector3;
     private _test: Test | undefined;
-    private callback: () => void = () => {};
+
+    private _resolve: (value: Test | PromiseLike<Test>) => void;
+    public readonly ready = new Promise<Test>(resolve => {
+        this._resolve = resolve;
+    });
 
     get testLocation(): Vector3 {
         return this._testLocation;
@@ -29,7 +35,7 @@ export class TestManager {
             world.gameRules.doMobSpawning = doMobSpawning;
 
             this._test = test;
-            this.callback();
+            this._resolve(test);
 
             console.log('[模拟玩家] 初始化完成，输入“假人创建”或“ffpp”');
         })
@@ -37,7 +43,24 @@ export class TestManager {
             .structureName('xboyMinemcSIM:void');
     }
 
-    onRegistered(callback: () => void): void {
-        this.callback = callback;
+    private generateTestPosition(): Vector3 {
+        const z = 11451400 + Math.floor(Math.random() * 114514 * 19);
+        return {
+            x: 15000000,
+            y: 256,
+            z
+        };
+    }
+
+    initialize(): void {
+        const { x, y, z } = this.generateTestPosition();
+
+        system.run(() => {
+            try {
+                overworld.runCommand(`execute positioned ${x} ${y} ${z} run gametest run 我是云梦:假人`);
+            } catch (e) {
+                world.sendMessage('[模拟玩家] 报错了，我也不知道为什么' + e);
+            }
+        });
     }
 }
