+++ src/main.ts
@@ -10,14 +10,15 @@ import { TestManager } from './core/test/manager';
 export const simulatedPlayerManager=new SimulatedPlayerManager();
 
 export const testManager = new TestManager()
-testManager.onRegistered(()=>{
-    simulatedPlayerManager.test=testManager.test
-})
+testManager.ready.then(test => {
+    simulatedPlayerManager.test = test;
+});
 testManager.registerTest();
 
-world.afterEvents.worldLoad.subscribe(()=>{
+world.afterEvents.worldLoad.subscribe(() => {
     simulatedPlayerManager.initialize();
-})
+    testManager.initialize();
+});
 
 const broadcast = () => {
     world.sendMessage('[模拟玩家] 初始化完成，输入“假人创建”或“ffpp”');
